Envelope CC + remove unused imports
